AI_Security_Architecture_Analyser/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
google-genai==1.68.0
chromadb>=1.5.5
sentence-transformers>=5.3.0